Move raincloud base entity to separate module
