+++ homeassistant/components/raincloud/binary_sensor.py
@@ -16,8 +16,8 @@ import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import RainCloudEntity
 from .const import DATA_RAINCLOUD, ICON_MAP
+from .entity import RainCloudEntity
 
 _LOGGER = logging.getLogger(__name__)
