+++ homeassistant/components/raincloud/sensor.py
@@ -17,8 +17,8 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.icon import icon_for_battery_level
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import RainCloudEntity
 from .const import DATA_RAINCLOUD, ICON_MAP
+from .entity import RainCloudEntity
 
 _LOGGER = logging.getLogger(__name__)
